+++ src/detection/gpu/gpu_apple.c
@@ -9,20 +9,26 @@
 static double detectGpuTemp(const FFstrbuf* gpuName)
 {
     double result = 0;
-    const char* error;
+    const char* error = NULL;
 
-    if(ffStrbufStartsWithS(gpuName, "Apple M1"))
-        error = ffDetectCoreTemps(FF_TEMP_GPU_M1X, &result);
-    else if(ffStrbufStartsWithS(gpuName, "Apple M2"))
-        error = ffDetectCoreTemps(FF_TEMP_GPU_M2X, &result);
-    else if(ffStrbufStartsWithS(gpuName, "Intel"))
+    if (ffStrbufStartsWithS(gpuName, "Apple M"))
+    {
+        switch (strtol(gpuName->chars + strlen("Apple M"), NULL, 10))
+        {
+            case 1: error = ffDetectCoreTemps(FF_TEMP_GPU_M1X, &result); break;
+            case 2: error = ffDetectCoreTemps(FF_TEMP_GPU_M2X, &result); break;
+            case 3: error = ffDetectCoreTemps(FF_TEMP_GPU_M3X, &result); break;
+            default: error = "Unsupported Apple Silicon GPU";
+        }
+    }
+    else if (ffStrbufStartsWithS(gpuName, "Intel"))
         error = ffDetectCoreTemps(FF_TEMP_GPU_INTEL, &result);
-    else if(ffStrbufStartsWithS(gpuName, "Radeon") || ffStrbufStartsWithS(gpuName, "AMD"))
+    else if (ffStrbufStartsWithS(gpuName, "Radeon") || ffStrbufStartsWithS(gpuName, "AMD"))
         error = ffDetectCoreTemps(FF_TEMP_GPU_AMD, &result);
     else
         error = ffDetectCoreTemps(FF_TEMP_GPU_UNKNOWN, &result);
 
-    if(error)
+    if (error)
         return FF_GPU_TEMP_UNSET;
 
     return result;